backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
fastapi==0.128.1
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0

# Database
sqlalchemy==2.0.41
alembic==1.16.1
psycopg2-binary==2.9.10

# Auth
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0

# Config
python-dotenv==1.1.0

# Optimization
scikit-optimize==0.10.2

# HTTP client (Ollama)
httpx==0.28.1

# Email validation
email-validator==2.2.0
